Switch Fisher to float64

+++ many/stats/categorical_categorical.py
@@ -89,7 +89,7 @@ def mat_fisher_naive(a_mat, b_mat, melt: bool, pseudocount=0, pbar=False):
     fishers: -log10 p-values of Fisher exact test
     """
 
-    a_mat, b_mat = precheck_align(a_mat, b_mat, np.int64, np.int64)
+    a_mat, b_mat = precheck_align(a_mat, b_mat, np.float64, np.float64)
 
     a_names = a_mat.columns
     b_names = b_mat.columns
@@ -205,7 +205,7 @@ def mat_fisher(a_mat, b_mat, melt: bool, pseudocount=0):
     fishers: -log10 p-values of Fisher exact test
     """
 
-    a_mat, b_mat = precheck_align(a_mat, b_mat, np.int64, np.int64)
+    a_mat, b_mat = precheck_align(a_mat, b_mat, np.float64, np.float64)
 
     a_names = a_mat.columns
     b_names = b_mat.columns
@@ -231,7 +231,7 @@ def mat_fisher(a_mat, b_mat, melt: bool, pseudocount=0):
     aB = np.dot(A_neg.T, B_pos) + pseudocount
     ab = np.dot(A_neg.T, B_neg) + pseudocount
 
-    comb = np.stack([AB, Ab, aB, ab]).astype(np.int64)
+    comb = np.stack([AB, Ab, aB, ab]).astype(np.float64)
 
     pvals = np.apply_along_axis(fisher_arr, 0, comb)
     pvals = pd.DataFrame(pvals, index=a_names, columns=b_names)
@@ -281,7 +281,7 @@ def mat_fisher_nan(a_mat, b_mat, melt: bool, pseudocount=0):
     fishers: -log10 p-values of Fisher exact test
     """
 
-    a_mat, b_mat = precheck_align(a_mat, b_mat, np.int64, np.int64)
+    a_mat, b_mat = precheck_align(a_mat, b_mat, np.float64, np.float64)
 
     a_names = a_mat.columns
     b_names = b_mat.columns
@@ -299,10 +299,10 @@ def mat_fisher_nan(a_mat, b_mat, melt: bool, pseudocount=0):
     a_nan = np.isnan(a_mat)
     b_nan = np.isnan(b_mat)
 
-    a_pos = np.ma.array(a_pos, mask=a_nan, dtype=np.int64)
-    a_neg = np.ma.array(a_neg, mask=a_nan, dtype=np.int64)
-    b_pos = np.ma.array(b_pos, mask=b_nan, dtype=np.int64)
-    b_neg = np.ma.array(b_neg, mask=b_nan, dtype=np.int64)
+    a_pos = np.ma.array(a_pos, mask=a_nan, dtype=np.float64)
+    a_neg = np.ma.array(a_neg, mask=a_nan, dtype=np.float64)
+    b_pos = np.ma.array(b_pos, mask=b_nan, dtype=np.float64)
+    b_neg = np.ma.array(b_neg, mask=b_nan, dtype=np.float64)
 
     AB = np.ma.dot(a_pos.T, b_pos) + pseudocount
     Ab = np.ma.dot(a_pos.T, b_neg) + pseudocount
